mage-scheduler-plugin/pyproject.toml
[project]
name = "mage-scheduler"
version = "2.0.0"
description = "Task scheduler plugin for Mage Lab — schedule commands, cron tasks, and dependency chains with no external services required."
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Mage Lab Community Tools Contributors"},
]

dependencies = [
    "apscheduler>=3.10.0",
    "croniter>=1.4.0",
    "dateparser>=1.2.0",
    "fastapi>=0.115.0",
    "httpx>=0.27.0",
    "jinja2>=3.1.3",
    "mcp>=1.0.0",
    "psutil>=5.9.0",
    "python-multipart>=0.0.9",
    "sqlalchemy>=2.0.0",
    "uvicorn>=0.30.0",
    "tzdata>=2024.1; sys_platform == 'win32'",
]

[project.urls]
Repository = "https://github.com/mage-lab/mage-lab-community-tools"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["mage_scheduler", "."]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.0.0",
]
